requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
beautifulsoup4==4.12.2
behave==1.2.6
behavex==3.2.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cmake==3.25.2
colorama==0.4.6
comtypes==1.4.11 ; sys_platform == 'win32'
configobj==5.0.8
csscompressor==0.9.5
decorator==4.4.2
distro==1.9.0
elevenlabs==2.4.0
exceptiongroup==1.2.2
feedparser==6.0.11
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.4
jiter==0.10.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moviepy==1.0.3
numpy==1.24.4
oauthlib==3.3.1
openai==1.90.0
OpenAIAuth==3.0.0
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.3
pexels-api==1.0.1
pillow==11.2.1
playwright==1.52.0
proglog==0.1.12
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.3
pypiwin32==223 ; sys_platform == 'win32'
PySocks==1.7.1
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==310 ; sys_platform == 'win32'
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
revChatGPT==6.8.6
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
selenium==4.16.0
setuptools==58.1.0
sgmllib3k==1.0.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.7
tiktoken==0.9.0
tls-client==1.0.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yt-dlp==2025.6.9
